Modificacion intereses personales

+++ resources/views/interesesPersonales/index.blade.php
@@ -12,24 +12,25 @@
                         <thead class="text-light">
                         <tr>
                             <th scope="col" width="20%">Nombre</th>
-                            <th scope="col" width="20%">Editar</th>
-                            <th scope="col" width="20%">Eliminar</th>
+                            <th scope="col" width="20%">Tipo de operación</th>
+                            <th scope="col" width="20%">Opciones</th>
                         </thead>
                         <tbody class="text-uppercase">
                           @foreach ($intereses as $item)
                             <tr>
                               <td>{{$item->nombre}} {{$item->primer_apellido}} {{$item->segundo_apellido}}</td>
+                              <td>{{$item->tipoOperaciones->valor}}</td>
                               <td>
-                                  <a href="{{route("intereses_personales.edit",$item->id)}}" type="button"
-                                     class="btn btn-warning btn-sm ion ion-edit"></a>
-                              </td>
-                              <td>
+                                @if($item->tipo_operacion_id != 4) 
                                   {!! Form::open(['action' => ['InteresesController@destroy', $item->id], 'method'=>'DELETE']) !!}
                                   <div style="display: inline-block;">
-                                      <button
+                                    <a href="{{route("intereses_personales.edit",$item->id)}}" type="button"
+                                     class="btn btn-warning btn-sm ion ion-edit"></a>
+                                      <button type="button" data-enviado="{{$item->enviado}}"
                                           class="btn btn-danger btn-sm ion ion-android-delete btn-borrar"></button>
                                   </div>
                                   {!! Form::close() !!}
+                                @endif
                               </td>
                             </tr>
                           @endforeach
@@ -62,24 +63,57 @@
         </div>
     </div>
     </div>
+    <div class="modal" tabindex="-1" role="dialog" id="modal_baja">
+        <div class="modal-dialog" role="document">
+                <div class="modal-content">
+                    <div class="modal-header">
+                        <h5 class="modal-title">Baja</h5>
+                        <button type="button" class="close" data-dismiss="modal" aria-label="Close">
+                        <span aria-hidden="true">&times;</span>
+                        </button>
+                    </div>
+                    {!! Form::open(['action' => ['InteresesController@destroy', 1], 'method'=>'DELETE','id' => 'frmBorrar']) !!}
+                    <input name="id" type="hidden" id='id_registro'>
+                    <div class="modal-body">
+                        <div class="form-row">
+                            <div class="form-group col-md-12">
+                                <strong>Motivo de baja</strong>
+                                {!! Form::select('motivo_baja_id',$motivos, [] ,['class'=>'form-control text-uppercase','placeholder' => 'SELECCIONA UNA OPCIÓN' ,'id' => 'motivo_baja_id','required' => true]) !!}
+                                <span class="text-danger" style="font-size:150%"></span>
+                            </div>
+                        </div>
+                    </div>
+                    <div class="modal-footer">
+                        <button type="submit" class="btn btn-sm btn-submit text-light">Eliminar</button>
+                        <a class="btn btn-secondary" data-dismiss="modal">Close</a>
+                    </div>
+                    {!! Form::close() !!}
+            </div>
+        </div>
+    </div>
 @endsection
 @section('scripts')
     <script>
-        $('.btn-borrar').on('click', function (e) {
-            let that = this;
-            e.preventDefault();
-            Swal.fire({
-                title: '¿Está seguro?',
-                text: 'Al oprimir el botón de aceptar se eliminará el registro',
-                icon: 'warning',
-                showCancelButton: true,
-                cancelButtonText: 'Cancelar'
-            }).then((result) => {
-                if (result.isConfirmed){
-                    $(that).closest('form').submit();
+        $(document).on('click', '.btn-borrar', function (e) {
+                var id = $(this).data('id');
+                var enviado = $(this).data('enviado');
+                if(enviado){
+                    $("#id_registro").val(id);
+                    $("#modal_baja").modal("show");
+                } else {
+                  Swal.fire({
+                    title: '¿Está seguro?',
+                    text: 'Al oprimir el botón de aceptar se eliminará el registro',
+                    icon: 'warning',
+                    showCancelButton: true,
+                    cancelButtonText: 'Cancelar'
+                }).then((result) => {
+                    if (result.isConfirmed) {
+                        $(this).closest('form').submit();
+                    }
+                });  
                 }
             });
-        });
 
         $('.btn-ninguno').on('click', function (e) {
             let that = this;
